Add a user agent to requests made to Treeherder's API
In bug 1230222 we're encouraging consumers of Treeherder's API to set
user agent. Ideally the two definitions of the user agent in this PR
would be combined, however the current repo structure (lack of modules),
sys.path manipulation and mixture of requests and urllib2, means it
would involve a bit of refactoring, so this small bit of duplication
seems fine for now.