Refactor internals of _url_for_revision()
Due to a limitation on Treeherder we cannot request more than 2000 jobs on a revision.
Unfortunately, we hit this limit very easily on mozilla-central where coalescing happens less
and we even have several hundred jobs more because of nightly jobs.

This change is a no-op change in preparation for using the TaskCluster index instead of
Treeherder to query about jobs.