Add support to query TaskCluster's index
When we want to query Treeherder jobs we have a 2000 jobs limitation.
With TaskCluster's index we can query for the task ID of specific task
by using routes that will lead us to the right job.

In this change we use the index to find the latest nightly build for a
specific repo and platform.